# Conflicts:
#	TutorOnline/src/main/java/com/pmi/tutor/controller/UserController.java

+++ TutorOnline/src/main/java/com/pmi/tutor/domain/Category.java
@@ -18,7 +18,7 @@ public class Category {
 	@Id
 	@GeneratedValue
 	@Column(name = "id")
-	private Integer id;
+	private Long id;
 
 	@Column(name = "name")
 	private String name;
@@ -26,11 +26,13 @@ public class Category {
 	@OneToMany(cascade = { CascadeType.ALL }, mappedBy = "category", fetch = FetchType.LAZY)
 	private Set<Subject> subjects = new HashSet<>();
 
-	public Integer getId() {
+
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
